+++ common/src/main/java/net/wizards/config/LootConfig.java
@@ -6,11 +6,11 @@ import java.util.List;
 public class LootConfig {
     public HashMap<String, ItemGroup> item_groups = new HashMap<>();
     public static class ItemGroup { public ItemGroup() { }
-        public List<String> items = List.of();
+        public List<String> ids = List.of();
         public int weight = 1;
 
-        public ItemGroup(List<String> items, int weight) {
-            this.items = items;
+        public ItemGroup(List<String> ids, int weight) {
+            this.ids = ids;
             this.weight = weight;
         }
     }